Tokenizer improvements, build improvements
* CBTextTokenizer is now thread-safe so a single instance can be used
  in the index. Now kept as an instance variable so the client can
  configure it.
* Added an OS X static library target.
* Updated ForestDB.
* Fixed the iOS build.
* Fixed some unit-test issues and compiler warnings.

+++ CBForest/CBCollatable.m
@@ -328,7 +328,7 @@ static uint8_t* getCharPriorityMap(void) {
         static const char* const kInverseMap = "\t\n\r `^_-,;:!?.'\"()[]{}@*/\\&#%+<=>|~$0123456789aAbBcCdDeEfFgGhHiIjJkKlLmMnNoOpPqQrRsStTuUvVwWxXyYzZ";
         uint8_t priority = 1;
         for (int i=0; i<strlen(kInverseMap); i++)
-            kCharPriority[kInverseMap[i]] = priority++;
+            kCharPriority[(uint8_t)kInverseMap[i]] = priority++;
         for (int i=128; i<256; i++)
             kCharPriority[i] = (uint8_t)i;
     });
